requirements.txt
Django>=3.2,<4.0
djangorestframework
yfinance
celery
redis
django-celery-beat